[AUDIT] exhaustive PostgreSQL code review — 30 bugs fixed across 20 files
5 parallel agents audited every server route file. Fixes by category:

COUNT(*) bigint→string (8 files): auth.ts, pilotAuth.ts, contacts.ts x3,
  news.ts, documents.ts x2, projects.ts, procedures.ts — all cast to ::int
  and parsed with Number() to prevent string/number type confusion in
  pagination, admin guards, and cleanup logic.

Silent data loss (2 fixes):
  - contacts.ts POST: fullNameDisplay missing from INSERT — every new contact
    silently lost the fullNameDisplay value, always defaulting to 1
  - ai.ts: UPDATE aiSystemPrompt silently no-ops on fresh install (no row) —
    changed to INSERT ... ON CONFLICT DO UPDATE

Critical schema bugs in sites/helpers.ts:
  - Removed 3 ghost columns (windSpeedMinIdeal/Max/Ideal) that don't exist
    in DB — archive restore crashed with PG column-not-found errors
  - Added missing launchHeight2, landingHeight2, closurePillsMax — archive
    restore silently discarded these field values

Security: xc.ts /xc/zones/refresh had no auth guard — any user could trigger
  expensive external downloads repeatedly (DoS vector); requireAuth added

Unquoted camelCase column (shop.ts): admin_sessions.userId → "userId" in JOIN.
  isAdminRequest() always returned false — admin product visibility completely
  broken, hidden products never shown to admins.

PG type errors:
  - tidyhq.ts: WHERE id = $1 on SERIAL column → $1::int cast added
  - settings.ts: TIMESTAMPTZ + 'Z' string corruption — removed '+Z' suffix
  - safety.ts: boolean enabled passed to INTEGER column — explicit ternary cast

Missing rowCount checks (2 fixes): sites/crud.ts PUT, pages.ts PUT — both
  returned 200 success on updates to non-existent records.

Transactions: db.ts migrations now wrapped in BEGIN/COMMIT/ROLLBACK — partial
  migration no longer leaves DB in permanently broken state.

Other: pilotAuth.ts null guard on re-fetch after tracker update, || → ?? for
  sortOrder/TTL/coords to stop clobbering legitimate falsy values,
  closures.ts N+1 INSERT loop replaced with single unnest() bulk INSERT.

+++ server/routes/auth.ts
@@ -295,7 +295,7 @@ router.post("/users", requireAuth, asyncHandler(async (req, res) => {
 
 router.delete("/users/:id", requireAuth, asyncHandler(async (req, res) => {
   const adminCount = await queryOne<{ count: number }>(
-    `SELECT COUNT(*) as count FROM contacts WHERE "isAdmin" = 1`
+    `SELECT COUNT(*)::int as count FROM contacts WHERE "isAdmin" = 1`
   );
   if ((adminCount?.count ?? 0) <= 1) {
     return res.status(400).json({ error: "Cannot delete the last admin user" });
